Index1D: changed from LONG k to INT k

+++ lawa/methods/adaptive/datastructures/index.cpp
@@ -25,7 +25,8 @@ Index1D::Index1D(void)
 : j(0), k(0), xtype(XBSpline)
 {}
 
-Index1D::Index1D(int _j, long _k, XType _xtype)
+//Index1D::Index1D(int _j, long _k, XType _xtype)
+Index1D::Index1D(int _j, int _k, XType _xtype)
 : j(_j), k(_k), xtype(_xtype)
 {}
